Make flux-rs a lib crate
* Rename flux-attrs to flux-attrs-impl
* Add flux-attrs as a proc macro crate imported by flux-rs
