+++ lib/flux-attrs-impl/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "flux-attrs-impl"
+version = "0.1.0"
+
+edition.workspace = true
+
+[dependencies]
+proc-macro2 = "1"
+quote = "1"
+syn = { version = "2", features = ["full", "extra-traits", "visit-mut"] }
+
+[lints]
+workspace = true
