+++ lib/flux-attrs/src/lib.rs
@@ -1,145 +1,183 @@
-mod ast;
-mod extern_spec;
+#[cfg(not(flux_sysroot))]
+use attr_dummy as attr_impl;
+#[cfg(flux_sysroot)]
+use attr_sysroot as attr_impl;
+use proc_macro::TokenStream;
 
-use proc_macro2::{Ident, TokenStream, TokenTree};
-use quote::{quote, quote_spanned, ToTokens};
-use syn::{parse_quote, spanned::Spanned, Attribute, ItemEnum, ItemStruct};
-
-pub const FLUX_ATTRS: &[&str] = &[
-    "assoc",
-    "field",
-    "generics",
-    "invariant",
-    "opaque",
-    "refined_by",
-    "sig",
-    "trusted",
-    "trusted_impl",
-    "variant",
-    "should_fail",
-    "opts",
-];
-
-pub fn extern_spec(attr: TokenStream, tokens: TokenStream) -> TokenStream {
-    extern_spec::transform_extern_spec(attr, tokens).unwrap_or_else(|err| err.to_compile_error())
+#[proc_macro_attribute]
+pub fn alias(attr: TokenStream, tokens: TokenStream) -> TokenStream {
+    attr_impl::alias(attr, tokens)
 }
 
-pub fn flux_tool_item_attr(name: &str, attr: TokenStream, item: TokenStream) -> TokenStream {
-    // I don't really know what I'm doing here, but spanning the quote with the item's span seems
-    // to behave correctly.
-    let span = item.span();
-    let name = TokenTree::Ident(Ident::new(name, span));
-    if attr.is_empty() {
-        quote_spanned! {span=>
-            #[flux_tool::#name]
-            #item
-        }
-    } else {
-        quote_spanned! {span=>
-            #[flux_tool::#name(#attr)]
-            #item
-        }
-    }
+#[proc_macro_attribute]
+pub fn sig(attr: TokenStream, tokens: TokenStream) -> TokenStream {
+    attr_impl::sig(attr, tokens)
 }
 
-pub fn refined_by(attr: TokenStream, item: TokenStream) -> TokenStream {
-    let span = item.span();
-    let mut item = match syn::parse2::<syn::Item>(item) {
-        Ok(item) => item,
-        Err(err) => return err.to_compile_error(),
-    };
-
-    match &mut item {
-        syn::Item::Enum(item_enum) => refined_by_enum(item_enum),
-        syn::Item::Struct(item_struct) => refined_by_struct(item_struct),
-        _ => return syn::Error::new(span, "expected struct or enum").to_compile_error(),
-    }
-
-    if cfg!(flux_sysroot) {
-        quote_spanned! {span=>
-            #[flux_tool::refined_by(#attr)]
-            #item
-        }
-    } else {
-        item.to_token_stream()
-    }
+#[proc_macro_attribute]
+pub fn spec(attr: TokenStream, tokens: TokenStream) -> TokenStream {
+    attr_impl::spec(attr, tokens)
 }
 
-fn refined_by_enum(item_enum: &mut ItemEnum) {
-    for variant in &mut item_enum.variants {
-        flux_tool_attrs(&mut variant.attrs);
-    }
+#[proc_macro_attribute]
+pub fn qualifiers(attr: TokenStream, tokens: TokenStream) -> TokenStream {
+    attr_impl::qualifiers(attr, tokens)
 }
 
-fn refined_by_struct(item_struct: &mut ItemStruct) {
-    for field in &mut item_struct.fields {
-        flux_tool_attrs(&mut field.attrs);
-    }
+#[proc_macro_attribute]
+pub fn refined_by(attr: TokenStream, tokens: TokenStream) -> TokenStream {
+    attr_impl::refined_by(attr, tokens)
 }
 
-fn flux_tool_attrs(attrs: &mut Vec<Attribute>) {
-    if cfg!(flux_sysroot) {
-        for attr in attrs {
-            transform_flux_attr(attr);
-        }
-    } else {
-        attrs.retain(|attr| !is_flux_attr(attr));
-    }
+#[proc_macro_attribute]
+pub fn invariant(attr: TokenStream, tokens: TokenStream) -> TokenStream {
+    attr_impl::invariant(attr, tokens)
 }
 
-fn path_is_one_of(path: &syn::Path, idents: &[&str]) -> bool {
-    idents.iter().any(|ident| path.is_ident(ident))
+#[proc_macro_attribute]
+pub fn constant(attr: TokenStream, tokens: TokenStream) -> TokenStream {
+    attr_impl::constant(attr, tokens)
 }
 
-fn is_flux_attr(attr: &syn::Attribute) -> bool {
-    let path = attr.path();
-    if path.segments.len() >= 2 {
-        let ident = &path.segments[0].ident;
-        ident == "flux" || ident == "flux_rs"
-    } else {
-        path_is_one_of(path, FLUX_ATTRS)
-    }
+#[proc_macro_attribute]
+pub fn opaque(attr: TokenStream, tokens: TokenStream) -> TokenStream {
+    attr_impl::opaque(attr, tokens)
 }
 
-fn transform_flux_attr(attr: &mut syn::Attribute) {
-    let path = path_of_attr_mut(attr);
-    if path.leading_colon.is_some() {
-        return;
-    }
-    if path.segments.len() >= 2 {
-        let ident = &mut path.segments[0].ident;
-        if ident == "flux" || ident == "flux_rs" {
-            *ident = Ident::new("flux_tool", ident.span());
-        }
-        return;
-    } else if path_is_one_of(path, FLUX_ATTRS) {
-        *path = parse_quote!(flux_tool::#path);
-    }
+#[proc_macro_attribute]
+pub fn opts(attr: TokenStream, tokens: TokenStream) -> TokenStream {
+    attr_impl::opts(attr, tokens)
 }
 
-fn path_of_attr_mut(attr: &mut Attribute) -> &mut syn::Path {
-    match &mut attr.meta {
-        syn::Meta::Path(path) => path,
-        syn::Meta::List(metalist) => &mut metalist.path,
-        syn::Meta::NameValue(namevalue) => &mut namevalue.path,
-    }
+#[proc_macro_attribute]
+pub fn trusted(attr: TokenStream, tokens: TokenStream) -> TokenStream {
+    attr_impl::trusted(attr, tokens)
 }
 
+#[proc_macro_attribute]
+pub fn trusted_impl(attr: TokenStream, tokens: TokenStream) -> TokenStream {
+    attr_impl::trusted_impl(attr, tokens)
+}
+
+#[proc_macro_attribute]
+pub fn generics(attr: TokenStream, tokens: TokenStream) -> TokenStream {
+    attr_impl::generics(attr, tokens)
+}
+
+#[proc_macro_attribute]
+pub fn assoc(attr: TokenStream, tokens: TokenStream) -> TokenStream {
+    attr_impl::assoc(attr, tokens)
+}
+
+#[proc_macro]
 pub fn flux(tokens: TokenStream) -> TokenStream {
-    syn::parse2::<ast::Items>(tokens)
-        .map_or_else(|err| err.to_compile_error(), ToTokens::into_token_stream)
+    flux_attrs_impl::flux(tokens.into()).into()
 }
 
+#[proc_macro]
 pub fn defs(tokens: TokenStream) -> TokenStream {
-    quote! {
-        #[flux::defs { #tokens }]
-        const _: () = {};
-    }
+    attr_impl::defs(tokens)
 }
 
-pub fn tokens_or_default<T: ToTokens + Default>(x: Option<&T>, tokens: &mut TokenStream) {
-    match x {
-        Some(t) => t.to_tokens(tokens),
-        None => T::default().to_tokens(tokens),
+#[proc_macro_attribute]
+pub fn extern_spec(attrs: TokenStream, tokens: TokenStream) -> TokenStream {
+    attr_impl::extern_spec(attrs, tokens)
+}
+
+#[proc_macro_attribute]
+pub fn ignore(attrs: TokenStream, tokens: TokenStream) -> TokenStream {
+    attr_impl::ignore(attrs, tokens)
+}
+
+#[proc_macro_attribute]
+pub fn should_fail(attrs: TokenStream, tokens: TokenStream) -> TokenStream {
+    attr_impl::should_fail(attrs, tokens)
+}
+
+#[cfg(flux_sysroot)]
+mod attr_sysroot {
+    use super::*;
+
+    pub fn extern_spec(attr: TokenStream, tokens: TokenStream) -> TokenStream {
+        flux_attrs_impl::extern_spec(attr.into(), tokens.into()).into()
     }
+
+    pub fn refined_by(attr: TokenStream, item: TokenStream) -> TokenStream {
+        flux_attrs_impl::refined_by(attr.into(), item.into()).into()
+    }
+
+    pub fn defs(tokens: TokenStream) -> TokenStream {
+        flux_attrs_impl::defs(tokens.into()).into()
+    }
+
+    macro_rules! flux_tool_attrs {
+        ($($name:ident),+ $(,)?) => {
+            $(
+            pub fn $name(attr: TokenStream, item: TokenStream) -> TokenStream {
+                flux_attrs_impl::flux_tool_item_attr(stringify!($name), attr.into(), item.into()).into()
+            }
+            )*
+        };
+    }
+
+    flux_tool_attrs!(
+        alias,
+        spec,
+        sig,
+        qualifiers,
+        constant,
+        invariant,
+        opaque,
+        opts,
+        trusted,
+        trusted_impl,
+        generics,
+        assoc,
+        ignore,
+        should_fail
+    );
+}
+
+#[cfg(not(flux_sysroot))]
+mod attr_dummy {
+    use super::*;
+
+    pub fn refined_by(attr: TokenStream, item: TokenStream) -> TokenStream {
+        flux_attrs_impl::refined_by(attr.into(), item.into()).into()
+    }
+
+    pub fn defs(_tokens: TokenStream) -> TokenStream {
+        TokenStream::new()
+    }
+
+    pub fn extern_spec(_attrs: TokenStream, _tokens: TokenStream) -> TokenStream {
+        TokenStream::new()
+    }
+
+    macro_rules! no_op {
+        ($($name:ident),+ $(,)?) => {
+            $(
+            pub fn $name(_attr: TokenStream, item: TokenStream) -> TokenStream {
+                item
+            }
+            )+
+        };
+    }
+
+    no_op!(
+        alias,
+        spec,
+        sig,
+        qualifiers,
+        invariant,
+        constant,
+        opaque,
+        opts,
+        trusted,
+        trusted_impl,
+        generics,
+        assoc,
+        ignore,
+        should_fail,
+    );
 }
