+++ lib/flux-rs/Cargo.toml
@@ -4,9 +4,6 @@ version = "0.1.0"
 
 edition.workspace = true
 
-[lib]
-proc-macro = true
-
 [dependencies]
 flux-attrs = { path = "../flux-attrs", version = "0.1.0" }
